Added shorter int types, fixed float/double for sqlite.

+++ exposed-gradle/src/test/resources/databases/example.kt
@@ -1,5 +1,5 @@
 object Testrelation : Table() {
     val intcolumn: Column<Int> = integer("intColumn")
     val textcolumn: Column<String> = text("textColumn")
-    val floatcolumn: Column<Float> = float("floatColumn")
+    val floatcolumn: Column<Double> = double("floatColumn")
 }
